refactor: 마이페이지 내 공연 내역 관련 컴포넌트 구조 개선
